Add files via upload
This is what I have so far. It runs but I am reading trash values from the initial call of R1.getx() and R1.gety(). I am not sure why, as they were initialized with (2,7) and I don't think that the moveTo function would work the way I have written it if it couldn't read those values.